Tie sorting GUI into sort methods
This commit along with the previous two implement a buggy way to
sort a deck based on different keys. In addition to the default
method of sorting by name, you can now also sort by price.

Bugs:
* Card categories are not sorted correctly when sorting by price
* Cards are sorted by name after insertion, regardless of the
  selected sorting method
* "Group by" functionality is not implemented

I'm unhappy with a few things in addition to the above. It would
be nice if I could get sorting to work by clicking column names
instead of using a dropdown. And I don't like the far-reaching
changes to method declarations needed to pass SortMethod around.
Rather than pass SortMethod through the various sort() methods, it
would probably be better to add it as a field to the card list or
somesuch, then reference that field internally in the relevant
sort() methods.

I will probably be creating another feature branch to try to re-
implement this soon.