Add code for different sorting methods
Give DeckLists some concept of alternate sorting strategies. This
is implemented by adding a DeckSort::SortMethod enum, and adding a
field of that type to InnerDecklistNodes. When a node's compare()
function is called, it checks the root node's SortMethod value to
determine how it should compare itself to other nodes.